Request can be affinitized to destinations (#174)
Request can be affinitized to destinations by an affinity key extracted from cookie or custom header. Affinity is active only when load balancing is enabled and it can be configured for each backend independently.
One request can have affinity to a single destination or to a set. It enables scenarios where a session needs to be affinitized to a pool of load-balanced destinations.
Affinity failures are handled by a policy configured per backend which has full access to `HttpContext` and decides whether the request processing can continue or not.

Fixes #45